Compiling: Errors, Warnings, and multiple outputs
One of the more frustrating things about PCT as a HLL developer was
always the extra verbose and non-informative messages that occurred
because an exception was thrown.  So include error and warning
handling in the system from step one.

Also, a single linear set of stages doesn't actually reflect how a
compiler works, so add some thoughts on the matter.