Death to POST, long live CFGs
Most of the operations currently done on POST can be sanely handled by
either a higher-level AST or on the CFGs.  I've also become more
convinced about the need of CFGs for reasonable optimization and useful
algorithms (like advanced register allocation).

So this commit removes the idea of an opcode tree in favor of control
flow graphs.  It may be that the AST layer may gain specialty stages or
nodes to deal with more low-level operations, but I no longer see a
need to deal with it as a completely separate layer.